requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.28.47
botocore==1.31.47
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
cmake==3.27.2
constantly==15.1.0
cryptography==41.0.3
Cython==3.0.2
daphne==4.0.0
Deprecated==1.2.14
Django==4.2.4
django-admin-rangefilter==0.11.1
django-cron==0.6.0
django-crontab==0.7.1
django-environ==0.10.0
django-filter==23.2
django-storages==1.14
djangorestframework==3.14.0
drf-yasg==1.21.7
environ==1.0
firebase-admin==6.2.0
gcloud==0.18.3
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
install==1.3.5
jmespath==1.0.1
jws==0.1.3
msgpack==1.0.5
Naked==0.1.32
oauth2client==4.1.3
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
psycopg==3.1.10
psycopg2-binary
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rollbar==0.16.3
rsa==4.9
s3transfer==0.6.2
service-identity==23.1.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
sseclient==0.0.27
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
zope.interface==6.0
